Move intent template code
