+++ corehq/apps/app_manager/views/form_designer.py
@@ -1,6 +1,5 @@
 import json
 import logging
-import os
 
 from django.utils.translation import ugettext as _
 from couchdbkit.exceptions import ResourceConflict
@@ -9,7 +8,6 @@ from django.shortcuts import render
 from django.views.decorators.http import require_GET
 from django.conf import settings
 from django.contrib import messages
-import yaml
 from corehq.apps.app_manager.views.apps import get_apps_base_context
 
 from corehq.apps.app_manager.views.utils import back_to_main, bail
@@ -18,6 +16,7 @@ from corehq.apps.accounting.utils import domain_has_privilege
 from corehq.apps.app_manager.util import (
     get_casedb_schema,
     get_session_schema,
+    app_callout_templates,
 )
 from corehq.apps.fixtures.fixturegenerators import item_lists_by_domain
 from corehq.apps.app_manager.dbaccessors import get_app
@@ -93,7 +92,7 @@ def form_designer(request, domain, app_id, module_id=None, form_id=None,
         'sessionid': request.COOKIES.get('sessionid'),
         'features': vellum_features,
         'plugins': vellum_plugins,
-        'app_callout_templates': next(_app_callout_templates),
+        'app_callout_templates': next(app_callout_templates),
     })
     return render(request, 'app_manager/form_designer.html', context)
 
@@ -169,27 +168,6 @@ def get_data_schema(request, domain, app_id=None, form_unique_id=None):
     if "pretty" in request.GET:
         kw["indent"] = 2
     return HttpResponse(json.dumps(data, **kw))
-
-
-def _app_callout_templates():
-    """Load app callout templates from config file on disk
-
-    Generator function defers file access until needed, acts like a
-    constant thereafter.
-    """
-    path = os.path.join(
-        os.path.dirname(__file__),
-        '..', 'static', 'app_manager', 'json', 'vellum-app-callout-templates.yaml'
-    )
-    if os.path.exists(path):
-        with open(path) as f:
-            data = yaml.load(f)
-    else:
-        logger.info("not found: %s", path)
-        data = []
-    while True:
-        yield data
-_app_callout_templates = _app_callout_templates()
 
 
 @require_can_edit_apps
